% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATMVolToSABRAlpha.R
\name{ATMVolToSABRAlpha}
\alias{ATMVolToSABRAlpha}
\title{ATM Vol to SABR Alpha calibration}
\usage{
ATMVolToSABRAlpha(F0, ATMVol, tex, Beta, Rho, Nu)
}
\arguments{
\item{F0}{Current forward rate}

\item{ATMVol}{Current BLACK ATM vol in the market corresponding to F0}

\item{tex}{Time to expiry of option, measured in years}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data OR preset by user}

\item{Rho}{Correlation between SABR forward and diffusion processes}

\item{Nu}{Vol-of-vol for SABR diffusion process}
}
\value{
Single numeric value correspoding to the calibrated ATM SABR Alpha
}
\description{
Returns the ATM SABR Alpha corresponding to the ATM Black market vol, using root solving to get the smallest positive root from SABRALPHACUBIC
}
\examples{
### Simple example involving a 3M option on the 10Y swap rate with 50\% Beta assumed
ATMVolToSABRAlpha(F0 = 0.0266, ATMVol = 0.4084, tex = 0.25, Beta = 0.5, Rho = -0.0356,
 Nu = 1.0504)

}
