% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Black76Delta.R
\name{Black76Delta}
\alias{Black76Delta}
\title{Black-76 Closed Form Delta}
\usage{
Black76Delta(F0, K, Vol, tex, rfr, CallOrPut)
}
\arguments{
\item{F0}{Current forward rate}

\item{K}{Strike of the option}

\item{Vol}{Implied volatility, MUST BE IN LOGNORMAL TERMS, can be taken from SABRTOBLACK76}

\item{tex}{Time to expiry, in years, of the option}

\item{rfr}{Riskless rate, best taken as either the 10y or 30y government zero rate}

\item{CallOrPut}{Takes values of 'c' or 'p', and nothing else - determines whether you are pricing a call or put}
}
\value{
Standard Black-76 Delta of a European option on a forward contract
}
\description{
Closed-form solution for the Delta of a European call or put on a forward contract or rate,
using the Black-76 formulation. For swaptions, provides the simple Delta for a European-style
exercise payer or receiver swaption.
}
\examples{
Black76Delta(F0 = 0.0266, K = 0.0250, Vol = 0.4084, tex = 0.25, rfr = 0.02, CallOrPut = 'c')
}
