% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Black76Vega.R
\name{Black76Vega}
\alias{Black76Vega}
\title{Black-76 Closed-Form Vega}
\usage{
Black76Vega(F0, K, Vol, tex, rfr)
}
\arguments{
\item{F0}{Current forward rate}

\item{K}{Strike of the option}

\item{Vol}{Implied volatility, MUST BE IN LOGNORMAL TERMS, can be taken from SABRTOBLACK76}

\item{tex}{Time to expiry, in years, of the option}

\item{rfr}{Riskless rate, best taken as either the 10y or 30y government zero rate}
}
\value{
Standard Black-76 Vega of a European option on a forward contract
}
\description{
Closed-form solution for the Vega of a European option on a forward contract or rate,
using the Black-76 formulation. For swaptions, provides the simple Gamma for a
European-style exercise payer or receiver swaption.
}
\examples{
Black76Vega(F0 = 0.0266, K = 0.0250, Vol = 0.4084, tex = 0.25, rfr = 0.02)
}
