% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curves}
\alias{curves}
\title{Curve data for SABRFunctions package}
\format{
A data frame in tidy format which contains 66 rows of 3 variables
\describe{
\item{Term}{Swap rate maturity in years, e.g. 0.25 means 3M, 1 means 1Y, etc.}
\item{Type}{Text string that tells you whether you are looking at a SOFR (swap) or ZERO (discount) rate}
\item{Value}{Actual value of the corresponding swap or zero rate for the given tenor point}
}
}
\source{
{Created using Bloomberg YCRV function from May 23, 2022}
}
\usage{
curves
}
\description{
Bloomberg SOFR swap and zero curve from May 23, 2022
}
\examples{
library(ggplot2)
data('curves')
ggplot(data = curves, aes(x = Term, y = Value, colour = Type)) +
geom_line() +
geom_point() +
labs(x = 'Term (Years)', y = 'Rate (\%)',
      title = 'USD SOFR Swap and Zero Curves',
      subtitle = 'Fixed 30/360 vs SOFR Rates as of May 23 2022',
      caption = '(Source: Bloomberg)',
      colour = 'Rate Type')
}
\keyword{datasets}
