% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histratevoldata}
\alias{histratevoldata}
\title{Historical Rate and Lognormal Vol Data for Beta Estimation}
\format{
A data frame with 36 observations of 4 variables
\describe{
\item{Rate}{Text label describing the particular forward rate, e.g. "3M10Y"}
\item{Forward}{Numerical vector of actual forward rate values}
\item{Date}{Observation date of the forward rate. Only for information.}
\item{BlackVol}{Lognormal ATM implied volatility corresponding to forward rate}
}
}
\source{
{Created using Bloomberg SWPM - OV pricing function from May 23, 2022}
}
\usage{
histratevoldata
}
\description{
Bloomberg swap forward rate and Black-76 implied volatility from May 23, 2022. Used for
vignette only.
}
\examples{
data('histratevoldata')
str(histratevoldata)
}
\keyword{datasets}
