% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRAlphaCubic.R
\name{SABRAlphaCubic}
\alias{SABRAlphaCubic}
\title{SABR Alpha Cubic Form for ATM Vols}
\usage{
SABRAlphaCubic(x, F0, ATMVol, tex, Beta, Rho, Nu)
}
\arguments{
\item{x}{The current value of SABR Alpha}

\item{F0}{Current value of forward rate}

\item{ATMVol}{Current value of ATM market volatility corresponding to F0}

\item{tex}{Time to expiry of the option, in years}

\item{Beta}{Calibrated (or user-selected) Beta parameter defining shape of the forward curve}

\item{Rho}{Calibrated correlation coefficient between forward and volatility Wiener processes}

\item{Nu}{Calibrated vol-of-vol for stochastic volatility process}
}
\value{
Single numeric value correspoding to the ATM Black-76-equivalent volatility for the given Alpha
}
\description{
Anonymous function to calculate the SABR Alpha given market inputs, used in calibration processes
}
\examples{
### Simple example involving a 3M option on the 10Y swap rate with 50\% Beta assumed
SABRAlphaCubic(x = 0.0651, F0 = 0.0266, ATMVol = 0.4084, tex = 0.25, Beta = 0.5, Rho = -0.0356,
 Nu = 1.0504)
}
