% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRDelta.R
\name{SABRDelta}
\alias{SABRDelta}
\title{SABR First-Order Risk with Backbone Correction}
\usage{
SABRDelta(F0, K, tex, rfr, CallOrPut, Alpha, Beta, Rho, Nu)
}
\arguments{
\item{F0}{Current forward price}

\item{K}{Strike of the option}

\item{tex}{Time to expiry, expressed in years, of the option}

\item{rfr}{Riskless rate, best taken as either the 10y or 30y government zero rate}

\item{CallOrPut}{Takes values of 'c' or 'p', and nothing else - determines whether you are pricing a call or a put}

\item{Alpha}{Diffusion parameter in SABR scheme, calibrated as a result of using one of two methods}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data OR preset by user}

\item{Rho}{Correlation between SABR forward and diffusion processes}

\item{Nu}{Vol-of-vol for SABR diffusion process}
}
\value{
Black-76-equivalent Delta of the option
}
\description{
Calculates the risk due to changes in the value of the underlying forward rate using a mixture
of analytical and numerical estimation methods, as outlined in 'Managing Smile Risk', p12.
Assumes that all parameters provided are ALREADY CALIBRATED, i.e. does not attempt to run any
calibrations first. Using the calibrated parameters, first calculates the Black-76-equivalent
volatility, then calculates the Delta risk of the option given the input parameters using
the analytical forms for Black-76 Delta and Vega.
}
\examples{
SABRDelta(F0 = 0.0266, K = 0.0250, tex = 0.25, rfr = 0.02, CallOrPut = 'c',
Alpha = 0.0651, Beta = 0.5, Rho = -0.0356, Nu = 1.0504)
}
