% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabrfitteddata}
\alias{sabrfitteddata}
\title{Results of calibration runs for SABRFunctions package}
\format{
A data frame with 168 rows of 4 variables:
\describe{
\item{Point}{Name of forward rate, used for filtering}
\item{Method}{Factor detailing what the data actually is - market quote, ATM SABR, Full SABR}
\item{Strike}{Strike rate of various options traded in the market}
\item{Value}{Implied lognormal volatility of option traded at that strike}
}
}
\source{
{Results from calibration runs from June 6, 2022}
}
\usage{
sabrfitteddata
}
\description{
Consolidated market volatilities, ATM SABR Black-76 vols, and Full SABR Black-76 vols, organised
by Strike and Method for easy subsetting and filtering.
}
\examples{
data('sabrfitteddata')  ###Lazy loading
}
\keyword{datasets}
