% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRFullCalib.R
\name{SABRFullCalib}
\alias{SABRFullCalib}
\title{Full Calibration of SABR Parameters Against Market Vols}
\usage{
SABRFullCalib(
  F0,
  Strikes,
  MarketVols,
  tex,
  Beta,
  guess_Alpha,
  guess_Rho,
  guess_Nu
)
}
\arguments{
\item{F0}{Current forward rate}

\item{Strikes}{VECTOR of strike prices}

\item{MarketVols}{VECTOR of LOGNORMAL (i.e. Black-76) market-quoted implied volatilities}

\item{tex}{Time to expiry of option, measured in years}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data OR preset by user}

\item{guess_Alpha}{Initial user-defined guess of Alpha value, MUST be non-zero}

\item{guess_Rho}{Initial user-defined guess of Rho value, MUST be bounded between -1 and 1}

\item{guess_Nu}{Initial user-defined guess of Nu value, MUST be non-zero}
}
\value{
List of outputs from the constrOptim function that includes the parameters for calibrated
Alpha/Rho/Nu
}
\description{
Calibrates Alpha Rho and Nu such that sum of square errors between Black-76-equivalent SABR vols
and market observed vols are minimised. For a given Alpha, Rho, and Nu, calculates the resulting
Black-76 vol, then takes SSE. Minimses SSE function using non-linear optimisation.
}
\examples{
data(sabrcalibdata)
calibrun <- sabrcalibdata[sabrcalibdata$Point == '3M10Y',]
SABRFullCalib(F0 = 0.0266, Strikes = calibrun$Strike, MarketVols = calibrun$BlackVol,
tex = 0.25, Beta = 0.5, guess_Alpha = 0.05, guess_Rho = 0.05, guess_Nu = 0.7)
}
