% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRGamma.R
\name{SABRGamma}
\alias{SABRGamma}
\title{SABR Second-Order Risk with Backbone Correction}
\usage{
SABRGamma(F0, K, tex, rfr, Alpha, Beta, Rho, Nu)
}
\arguments{
\item{F0}{Current forward price}

\item{K}{Strike of the option}

\item{tex}{Time to expiry, expressed in years, of the option}

\item{rfr}{Riskless rate, best taken as either the 10y or 30y government zero rate}

\item{Alpha}{Diffusion parameter in SABR scheme, calibrated as a result of using one of two methods}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data OR preset by user}

\item{Rho}{Correlation between SABR forward and diffusion processes}

\item{Nu}{Vol-of-vol for SABR diffusion process}
}
\value{
Corrected second-order risk against changes in the underlying forward rate
}
\description{
Calculates the second-order risk due to changes in the value of the underlying forward rate
using a mixture of analytical and numerical estimation methods. Assumes that all parameters
provided are ALREADY CALIBRATED, i.e. does not attempt to run any calibrations first. Using
the calibrated parameters, first calculates the Black-76-equivalent volatility, then calculates
the Delta risk of the option given the input parameters using the analytical forms for Black-76
Delta and Vega.
}
\examples{
SABRGamma(F0 = 0.0266, K = 0.0250, tex = 0.25, rfr = 0.02, Alpha = 0.0651, Beta = 0.5,
Rho = -0.0356, Nu = 1.0504)
}
