% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRParamLinearBump.R
\name{SABRParamLinearBump}
\alias{SABRParamLinearBump}
\title{Black-76-Equivalent Vols from Parameter Bumps for SABR Inputs}
\usage{
SABRParamLinearBump(
  F0,
  K,
  tex,
  Alpha,
  Beta,
  Rho,
  Nu,
  bump_param,
  bumpsize = 1/20000,
  bumpdir = "up"
)
}
\arguments{
\item{F0}{Current forward price}

\item{K}{Strike of the option}

\item{tex}{Time to expiry, expressed in years, of the option}

\item{Alpha}{Calibrated SABR Alpha value from either ATM or FULL calibration methods}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data OR preset by user}

\item{Rho}{Correlation between SABR forward and diffusion processes}

\item{Nu}{Vol-of-vol for SABR diffusion process}

\item{bump_param}{Specifies which parameter you want to bump. Must be one of the inputs listed above.}

\item{bumpsize}{Actual size of the bump, defaults to 0.5bps, but can be increased as required}

\item{bumpdir}{Allows for 'up' or 'dn' bumps}
}
\value{
Revised Black-76-equivalent SABR implied volatility
}
\description{
Calculates a one-sided bump of 0.5bp by default (user can change the bump size) for the specific
parameter that the  user wants to alter, for the Black-76-equivalent volatility. Uses a switch
function to figure out which parameter to bump, and then runs the actual shift.
}
\examples{
SABRParamLinearBump(F0 = 0.0266, K = 0.0250, tex = 0.25, Alpha = 0.0651, Beta = 0.5,
Rho = -0.0356, Nu = 1.0504, bump_param = 'Alpha', bumpdir = 'dn')
}
