% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRtoBlack76.R
\name{SABRtoBlack76}
\alias{SABRtoBlack76}
\title{Convert from SABR to Black-76 Volatility}
\usage{
SABRtoBlack76(F0, K, tex, Alpha, Beta, Rho, Nu)
}
\arguments{
\item{F0}{Current forward rate}

\item{K}{Strike rate of the option}

\item{tex}{Time ot expiry of the option, measured in years}

\item{Alpha}{Diffusion parameter in the SABR scheme, calibrated as a result of using one of two methods}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data, OR preset by user}

\item{Rho}{Correlation between SABR forward and diffusion processes}

\item{Nu}{Vol-of-vol parameter for SABR diffusion process}
}
\value{
Black-76-equivalent volatility that can then be plugged into usual Black-76 closed-form option price
}
\description{
Returns the Black-76 EQUIVALENT volatility after calibrating for SABR Alpha, Beta, Rho, and Nu, using the formulation found in
Eqn 2.17 of 'Managing Smile Risk' by Hagan et al, 2002. Base function for all other calibrations
}
\examples{
### Example for ITM option
SABRtoBlack76(F0 = 0.0266, K = 0.0250, tex = 0.25, Alpha = 0.0651, Beta = 0.5, Rho = -0.0356,
 Nu = 1.0504)

### Example for ATM option
SABRtoBlack76(F0 = 0.0266, K = 0.0266, tex = 0.25, Alpha = 0.0651, Beta = 0.5, Rho = -0.0356,
 Nu = 1.0504)

}
