% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRVolga.R
\name{SABRVolga}
\alias{SABRVolga}
\title{SABR First-Order Risk Against Vol-of-Vol Changes}
\usage{
SABRVolga(F0, K, tex, rfr, Alpha, Beta, Rho, Nu)
}
\arguments{
\item{F0}{Current forward rate}

\item{K}{Strike rate of the option}

\item{tex}{Time ot expiry of the option, measured in years}

\item{rfr}{Riskless rate, best taken as either the 10y or 30y government zero rate}

\item{Alpha}{Diffusion parameter in the SABR scheme, calibrated as a result of using one of two methods}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data, OR preset by user}

\item{Rho}{Correlation between SABR forward and diffusion processes}

\item{Nu}{Vol-of-vol parameter for SABR diffusion process}
}
\value{
First-order risk against changes to vol-of-vol
}
\description{
Calculates the first-order risk with respect to changes in the vol-of-vol parameter, using a
mixture of analytical and numerical estimation methods. Assumes that all parameters provided are
ALREADY CALIBRATED, i.e. does not attempt to run any calibrations first. Using the calibrated
parameters, first calculates the Black-76-equivalent volatility, then calculates the Vega risk of the option
given the input parameters using the analytical form for Black-76 Vega, and finally calculates the adjustment
required for the bump of SABR Nu.
}
\examples{
SABRVolga(F0 = 0.0266, K = 0.0250, tex = 0.25, rfr = 0.02, Alpha = 0.0651, Beta = 0.5,
Rho = -0.0356, Nu = 1.0504)
}
