% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRVolsFromATMCalib.R
\name{SABRVolsFromATMCalib}
\alias{SABRVolsFromATMCalib}
\title{ATM Calibration Wrapper Function for SABR Parameters}
\usage{
SABRVolsFromATMCalib(
  F0,
  ATMVol,
  Strikes,
  MarketVols,
  tex,
  Beta,
  guess_Rho,
  guess_Nu
)
}
\arguments{
\item{F0}{Current forward rate}

\item{ATMVol}{Current lognormal ATM volatility corresponding to F0 in the market}

\item{Strikes}{VECTOR of strike prices}

\item{MarketVols}{VECTOR of LOGNORMAL (i.e. Black-76) market-quoted implied volatilities}

\item{tex}{Time to expiry of option, measured in years}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data OR preset by user}

\item{guess_Rho}{Initial user-defined guess of Rho value, MUST be bounded between -1 and 1}

\item{guess_Nu}{Initial user-defined guess of Nu value, MUST be non-zero}
}
\value{
A list object containing each of the 4 calibrated parameters, plus a vector of
the input strikes, and a vector of the calibrated Black-76-equivalent volatilities.
}
\description{
Runs the complete calibration against market volatilities and strikes by calling the ATM
calibration method, SABRATMCALIB.
}
\examples{
data(sabrcalibdata)
calibrun <- sabrcalibdata[sabrcalibdata$Point == '3M10Y',]
SABRVolsFromATMCalib(F0 = 0.0266, ATMVol = 0.4084, Strikes = calibrun$Strike,
MarketVols = calibrun$BlackVol, tex = 0.25, Beta = 0.5, guess_Rho = 0.05, guess_Nu = 0.7)

}
