% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABRVolsFromFullCalib.R
\name{SABRVolsFromFullCalib}
\alias{SABRVolsFromFullCalib}
\title{Full Calibration Wrapper Function for SABR Parameters}
\usage{
SABRVolsFromFullCalib(
  F0,
  Strikes,
  MarketVols,
  tex,
  Beta,
  guess_Alpha,
  guess_Rho,
  guess_Nu
)
}
\arguments{
\item{F0}{Current forward rate}

\item{Strikes}{VECTOR of strike prices}

\item{MarketVols}{VECTOR of LOGNORMAL (i.e. Black-76) market-quoted implied volatilities}

\item{tex}{Time to expiry of option, measured in years}

\item{Beta}{Shape parameter of SABR schema, EITHER evaluated using historical data OR preset by user}

\item{guess_Alpha}{Initial user-defined guess of Alpha parameter, MUST be non-zero}

\item{guess_Rho}{Initial user-defined guess of Rho value, MUST be bounded between -1 and 1}

\item{guess_Nu}{Initial user-defined guess of Nu value, MUST be non-zero}
}
\value{
A list object containing each of the 4 calibrated parameters, plus a vector of
the input strikes, and a vector of the calibrated Black-76-equivalent volatilities.
}
\description{
Runs the complete calibration against market volatilities and strikes by calling the FULL
calibration method, SABRFULLCALIB.
}
\examples{
data(sabrcalibdata)
calibrun <- sabrcalibdata[sabrcalibdata$Point == '3M10Y',]
SABRVolsFromFullCalib(F0 = 0.0266, Strikes = calibrun$Strike, MarketVols = calibrun$BlackVol,
tex = 0.25, Beta = 0.5, guess_Alpha = 0.06, guess_Rho = 0.05, guess_Nu = 0.7)

}
